Move config struct to its own file
